move commonly used code of the webview extensions to the sourcepoint_unified_cmp package
