+++ packages/sourcepoint_unified_cmp_flutter_webview_extension/pubspec.yaml
@@ -12,7 +12,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  sourcepoint_unified_cmp_platform_interface: ^0.1.22
+  sourcepoint_unified_cmp: ^0.1.22
   webview_flutter: ^4.7.0
 
 dev_dependencies:
