+++ packages/sourcepoint_unified_cmp_flutter_webview_extension/test/sourcepoint_unified_cmp_flutter_webview_extension_test.dart
@@ -1,6 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:sourcepoint_unified_cmp_flutter_webview_extension/sourcepoint_unified_cmp_flutter_webview_extension.dart';
-import 'package:sourcepoint_unified_cmp_platform_interface/sourcepoint_unified_cmp_platform_interface.dart';
+import 'package:sourcepoint_unified_cmp/sourcepoint_unified_cmp.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
